# feat: use channel instead of mutex in scheduling predicates