**What type of PR is this?**

/kind cleanup
/sig scheduling
/assign @Huang-Wei @bsalamat 

**What this PR does / why we need it**:

Use channel and select statement instead of mutex in scheduling predicates.

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
NONE
```