@SataQiu 2019-07-04
Thanks @draveness 
In my view, if we're only care about the first error, using channel will increases unnecessary memory consumption. How about change it as :
```go
	catchError := func(err error) {
            if firstError == nil {
                lock.Lock()
		defer lock.Unlock()
                if firstError == nil {
                    firstError = err
                }
	    }
	}
```

 `topologyPairsCh` is OK for me.