@draveness 2019-07-04
> In my view, if we're only care about the first error, using channel will increases unnecessary memory consumption. How about change it as :

I think to save use channel and a bit more memory is better than using a mutex.