@draveness 2019-07-06
> Thanks, note that the compiler has nothing to do with what we are discussing here. By compressing the code, I am only talking about being able to see more lines of code in a single screen when reading/reviewing code.

I saw several of your PRs, which also add some newlines to the codebase, newline is part of our code and helps us to understand it more easily. I don't think to see more lines in one screen would help us understand the code more efficiently. Though it is a pro, this doesn't sound reasonable to me. 


And I see your point is to be consistent and follow the convention, but since there isn't a standard way to guide us "when and where we should put/remove a new line", and I noticed you are not asking to remove all the new line in the codebase, call it a convention is not good enough. 


```go

 	close(affinityPodsCh)
	close(antiAffinityPodsCh)
// why remove this but not the previous one.
```

The latest commit has removed several new lines that seem to make sense. And I really don't want to take much time on this kind of staff. Thank you for your reviews anyway.