@ahg-g 2019-07-06
> > Thanks, note that the compiler has nothing to do with what we are discussing here. By compressing the code, I am only talking about being able to see more lines of code in a single screen when reading/reviewing code.
> 
> I saw several of your PRs, which also add some newlines to the codebase, newline is part of our code and helps us to understand it more easily. I don't think to see several more lines in one screen would help us understand or review the code more efficiently. Though it could be a benefit, this doesn't sound reasonable to me.
> 
> And I see your point is to be consistent and follow the convention, but since there isn't a standard way to guide us "when and where we should put/remove a new line", and I noticed you are not asking to remove all the new line in the codebase, call it a convention is not good enough.
> 
> ```go
>  	close(affinityPodsCh)
> 	close(antiAffinityPodsCh)
> // why remove this but not the previous one.
> ```
> 
> Besides, I just posted a question on #k8s-code-organization channel, and we don't seem like to have this convention. https://kubernetes.slack.com/archives/CHGFYJVAN/p1562381796191600
> 
> > we just use gofmt to format code.

I am not arguing against adding empty lines in general, nor trying to suggest that there are rules where to add one and where not to, that just does not make sense, I am just making a general comment that by looking at the codebase one can see that it favors compactness. My original feedback was to remove a single line that looked a little odd compared to the rest of the code, the rest where in response to your ask to "point out which new lines you prefer to remove. ", so I provided my preference, that is all.

> 
> The latest commit has removed several new lines that seem to make sense 
> and I really don't want to take much time on this kind of staff. Thank you for your reviews anyway.

Indeed this back and forth wasn't the best use of one's time, it wasn't my intent to be picky...