@ahg-g 2019-07-08
I think this still does not address the concern that the current implementation may incur non trivial memory overhead (and possibly changes the performance profile). Can you provide benchmarking numbers for this change in this PR? otherwise we could have the non-error related changes in a separate PR with benchmarking numbers.