@ahg-g 2019-07-08
> > I think this still does not address the concern that the current implementation may incur non trivial memory overhead (and possibly changes the performance profile). Can you provide benchmarking numbers for this change in this PR? otherwise we could have the non-error related changes in a separate PR with benchmarking numbers.
> 
> **before**
> 
> ```go
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8         	    1000	  82154233 ns/op
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8         	    1000	  83304799 ns/op
> ```
> 
> **after**
> 
> ```
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8         	    1000	  84032711 ns/op
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8         	    1000	  83207072 ns/op
> ```
> 
> It does not show performance degradation from this benchmark.

Thanks for running the benchmark, in addition to the performance degradation it adds memory overhead. So I think we should split this PR to focus only on the error since that should have no impact, and leave the others as it is.