@draveness 2019-07-08
> Thanks for running the benchmark, in addition to the performance degradation it adds memory overhead. So I think we should have this PR focused only on the error since that should have no performance impact, and leave the others as it is.

I prefer to keep the current logic since this does not introduce performance degradation and have a better code structure.