@draveness 2019-07-08
> It shows 2% degradation, which even if we consider not significant, it also adds memory overhead.

I'm not agreeing with you on this, if you run it multiple times, I think you could see a performance improvement by several percent. The benchmark test is not stable, it could show a different result in the same environment.